wip: restructure matcher module
